requirements_pi.txt
ai-edge-litert==2.1.2
audioop-lts==0.2.2
audioread==3.1.0
backports.strenum==1.2.8
bleak==2.1.1
bluepy==1.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
dbus-fast==4.0.0
decorator==5.2.1
flatbuffers==25.12.19
idna==3.11
joblib==1.5.3
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
msgpack==1.1.2
numba==0.63.1
numpy==2.3.5
packaging==26.0
pandas==3.0.0
platformdirs==4.5.1
pooch==1.9.0
protobuf==6.33.5
psutil==7.2.2
pycparser==3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
setuptools==82.0.0
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
wheel==0.46.3
